Update PageHader.vue and Accounting Part
When the logout button was pressed in the page header, the session was cleared(), but the value of auth was not initialized, so the problem of opening other Menubars was solved by initializing through auth='; before reloading.
In addition, no matter which page you press the logout button, route to move it to the homepage screen.I did route.push ('/').

Finally, the accounting part processed the CRUD operation details.